Merge all stm32f4 demos in one repository
It turned out that keeping various stm32f4 examples (discovery and nucleo) in different
branches is not convenient since most part of code can be unified. Besides, all the
stm32-tests examples depend on libopencm3 which makes them even more similar.

This is the first iteration of merging all the stm32f4 demos in one repository. So far
build system is not elegant at all and need to be reconsidered. Directory tree is also not
finalized yet. Currently it is not clear what is to be done with shields.

Signed-off-by: matsi <matsi@example.org>