Workaround to improve current simple implementation of delay API
At the moment only libnrf24 needs delay API in this demo application. Note that library in
some cases needs delays of tens of microseconds. On the other hand, accuracy of delay is not
crucial: slightly longer delay is not going to be a problem, moreover longer is better in
most cases.

Current implementation of delay_[mu]s uses vTaskDelay.
Drawbacks and limitations:
  - usable only from tasks
  - libnrf24 needs up to ~10us delays => high RATE_HZ is needed

FIXME: think about better way to implement delay_xs functions.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>